+++ src/pages/test-report/cat-b/test-report.cat-b.page.scss
@@ -1,6 +1,6 @@
 $report-mode-border-width: 4px;
 $toolbar-height: 68px;
-page-test-report {
+test-report-cat-b-page {
 
   // ensure the content on the test report page doesn't grow with larger accessibility text
   ion-content {
