+++ src/pages/test-report/components/manoeuvres/manoeuvres.ts
@@ -1,6 +1,6 @@
 import { Manoeuvres } from '@dvsa/mes-test-schema/categories/B';
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
-import { OverlayCallback } from '../../test-report';
+import { OverlayCallback } from '../../cat-b/test-report.cat-b.page';
 import { StoreModel } from '../../../../shared/models/store.model';
 import { Store, select } from '@ngrx/store';
 import { getTestData } from '../../../../modules/tests/test-data/test-data.reducer';
