+++ src/pages/test-report/test-report.module.ts
@@ -1,34 +1,9 @@
 import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { TestReportPage } from './test-report';
-import { EffectsModule } from '@ngrx/effects';
-import { StoreModule } from '@ngrx/store';
-
-import { TestReportAnalyticsEffects } from './test-report.analytics.effects';
-import { testReportReducer } from './test-report.reducer';
-import { ComponentsModule } from '../../components/common/common-components.module';
-import { TestReportComponentsModule } from './components/test-report-components.module';
-import { TestReportValidatorProvider } from '../../providers/test-report-validator/test-report-validator';
-import { TestReportEffects } from './test-report.effects';
-import { TestResultProvider } from '../../providers/test-result/test-result';
+import { TestReportCatBPageModule } from './cat-b/test-report.cat-b.page.module';
 
 @NgModule({
-  declarations: [
-    TestReportPage,
-  ],
   imports: [
-    TestReportComponentsModule,
-    IonicPageModule.forChild(TestReportPage),
-    StoreModule.forFeature('testReport', testReportReducer),
-    EffectsModule.forFeature([
-      TestReportAnalyticsEffects,
-      TestReportEffects,
-    ]),
-    ComponentsModule,
-  ],
-  providers: [
-    TestReportValidatorProvider,
-    TestResultProvider,
+    TestReportCatBPageModule,
   ],
 })
 export class TestReportPageModule {}
